requirements.txt
torch
torchvision
torchaudio
diffusers
transformers
accelerate
safetensors
streamlit==1.28.0
requests==2.31.0
pillow==10.0.0